requirements.txt
numpy==1.21.2
opencv-python==4.5.5.62
Pillow==8.4.0
python-Levenshtein==0.12.2
matplotlib==3.5.0.